refactor: extract `_ShimmerInfoRow` into a new, reusable `ShimmerInfoRow` widget.
